Split up UsdBridgeUsdWriter into multiple parts
